+++ pvm-setup.iss
@@ -1,0 +1,71 @@
+; Script generated by the Inno Setup Script Wizard.
+; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
+
+#define MyAppName "PVM"
+#define MyAppVersion "1.2.0"
+#define MyAppPublisher "hjb.dev"
+#define MyAppURL "https://github.com/hjbdev/pvm"
+#define MyAppExeName "pvm.exe"
+
+[Setup]
+; NOTE: The value of AppId uniquely identifies this application. Do not use the same AppId value in installers for other applications.
+; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
+AppId={{8815C8A8-6961-4C32-A432-B762957813B8}}
+AppName={#MyAppName}
+AppVersion={#MyAppVersion}
+AppVerName={#MyAppName} {#MyAppVersion}
+AppPublisher={#MyAppPublisher}
+AppPublisherURL={#MyAppURL}
+AppSupportURL={#MyAppURL}
+AppUpdatesURL={#MyAppURL}
+DefaultDirName={%USERPROFILE}\.pvm
+UninstallDisplayIcon={app}\{#MyAppExeName}
+ArchitecturesAllowed=x64compatible
+ArchitecturesInstallIn64BitMode=x64compatible
+DefaultGroupName={#MyAppName}
+LicenseFile=LICENSE
+PrivilegesRequired=lowest
+PrivilegesRequiredOverridesAllowed=dialog
+OutputBaseFilename=pvm-setup
+SolidCompression=yes
+WizardStyle=modern
+
+[Languages]
+Name: "english"; MessagesFile: "compiler:Default.isl"
+
+[Files]
+Source: "{#SourcePath}\src\{#MyAppExeName}"; DestDir: "{app}"; Flags: ignoreversion
+Source: "{#SourcePath}\.pvm\*"; DestDir: "{app}"; Flags: ignoreversion recursesubdirs createallsubdirs
+; NOTE: Don't use "Flags: ignoreversion" on any shared system files
+
+[UninstallDelete]
+Type: filesandordirs; Name: "{app}"
+
+[Code]
+var
+  path: string;
+
+procedure InitializePath();
+begin
+  if RegQueryStringValue(HKEY_CURRENT_USER, 'Environment', 'Path', path) then
+  begin
+    if Pos(ExpandConstant('{app}'), path) = 0 then
+    begin
+      path := path + ';' + ExpandConstant('{app}') + '\bin';
+      StringChangeEx(path, ';;', ';', True);
+      RegWriteExpandStringValue(HKEY_CURRENT_USER, 'Environment', 'Path', path);
+    end;
+  end
+  else
+  begin
+    MsgBox('Unable to read PATH environment variable.', mbError, MB_OK);
+  end;
+end;
+
+procedure CurStepChanged(CurStep: TSetupStep);
+begin
+  if CurStep = ssPostInstall then
+  begin
+    InitializePath();
+  end;
+end;
